chore: move drill down menu item inside underlying data folder
