+++ packages/frontend/src/components/SimpleStatistic/BigNumberContextMenu.tsx
@@ -2,10 +2,9 @@ import { Menu, Position } from '@blueprintjs/core';
 import { MenuItem2, Popover2, Popover2Props } from '@blueprintjs/popover2';
 import { isField, isMetric, ResultRow } from '@lightdash/common';
 import { FC, useCallback, useMemo } from 'react';
-import { useParams } from 'react-router-dom';
 import { useExplore } from '../../hooks/useExplore';
-import DrillDownMenuItem from '../DrillDownMenuItem';
 import { useVisualizationContext } from '../LightdashVisualization/VisualizationProvider';
+import DrillDownMenuItem from '../UnderlyingData/DrillDownMenuItem';
 import { useUnderlyingDataContext } from '../UnderlyingData/UnderlyingDataProvider';
 
 interface BigNumberContextMenuProps {
@@ -15,7 +14,6 @@ interface BigNumberContextMenuProps {
 export const BigNumberContextMenu: FC<BigNumberContextMenuProps> = ({
     renderTarget,
 }) => {
-    const { projectUuid } = useParams<{ projectUuid: string }>();
     const { resultsData, bigNumberConfig } = useVisualizationContext();
     const { viewData, tableName } = useUnderlyingDataContext();
     const { data: explore } = useExplore(tableName);
@@ -63,7 +61,6 @@ export const BigNumberContextMenu: FC<BigNumberContextMenuProps> = ({
                         explore &&
                         resultsData && (
                             <DrillDownMenuItem
-                                projectUuid={projectUuid}
                                 row={resultsData.rows[0]}
                                 explore={explore}
                                 metricQuery={resultsData.metricQuery}
